+++ packages/theme/CHANGELOG.md
@@ -1,3 +1,25 @@
+## 3.14.9 (2026-06-01)
+
+This was a version bump only for theme to align it with other projects, there were no code changes.
+
+## 3.14.8 (2026-05-21)
+
+This was a version bump only for theme to align it with other projects, there were no code changes.
+
+## 3.14.7 (2026-05-12)
+
+This was a version bump only for theme to align it with other projects, there were no code changes.
+
+## 3.14.6 (2026-05-11)
+
+### 🩹 Fixes
+
+- **theme,components,select-styles:** swap icon and icon-sm size token values ([25a693d3ea2](https://github.com/migrationsverket/midas/commit/25a693d3ea2))
+
+## 3.14.5 (2026-05-11)
+
+This was a version bump only for theme to align it with other projects, there were no code changes.
+
 ## 3.14.4 (2026-05-08)
 
 This was a version bump only for theme to align it with other projects, there were no code changes.